Fix link to DOMException
One of the links to DOMException in the source used `[=DOMException=]` instead
of `{{DOMException}}` which made Bikeshed think that the link was to a similar
term definition in the WebCrypto API. The cross-references database was updated
to get rid of WebCrypto API terms that should not have been there, but that
means that the shorthand now needed to be fixed.

This update also adds wrapping brackets to other mentions of DOMException and
related terms (not required to fix generation with Bikeshed, just more
consistent).